+++ Bloch/Features/OnTurnEnd.cs
@@ -54,11 +54,11 @@ internal sealed class OnTurnEndManager : IWrappedActionHook
 		if (action is not TriggerAction triggerAction)
 			return true;
 
-		bool oldActionDisabled = triggerAction.Action.disabled;
+		var oldActionDisabled = triggerAction.Action.disabled;
 		triggerAction.Action.disabled = triggerAction.disabled;
 
 		var position = g.Push(rect: new()).rect.xy;
-		int initialX = (int)position.x;
+		var initialX = (int)position.x;
 
 		if (!dontDraw)
 			Draw.Sprite(ActionIcon.Sprite, position.x, position.y, color: action.disabled ? Colors.disabledIconTint : Colors.white);
@@ -94,7 +94,7 @@ internal sealed class OnTurnEndManager : IWrappedActionHook
 					Title = ModEntry.Instance.Localizations.Localize(["action", "OnTurnEnd", "name"]),
 					Description = ModEntry.Instance.Localizations.Localize(["action", "OnTurnEnd", "description"]),
 				},
-				..Action.GetTooltips(s)
+				.. Action.GetTooltips(s)
 			];
 
 		public override void Begin(G g, State s, Combat c)